ui: move header into component
